Galaxie Blog 4.019
Fixes bugs introduced by the recent ColdFusion update 14. The remote AJAX calls in the administrative site were failing due to HTML5 widgets producing extra arguments in the URL, especially with the jsGrids and the Kendo UI dropdowns. I will provide a more robust update soon but this should fix all of the new bugs introduced in the last week.